refactor(ui): merge /vulns into /findings, redirect orphan runtime routes, fix surface lock
- Merge the Findings implementation into its canonical /findings route and
  redirect the unlinked /vulns duplicate there, preserving query params.
- Redirect standalone /proxy and /gateway to their /runtime?tab= deep links;
  relocate the dashboards to sibling modules so /runtime still embeds them.
- Drop the wrong surface="agents" lock on the Findings empty state and make
  the proxy/gateway surface defs use plain route navHrefs like every other
  surface.
- Reuse canonical nav labels for the proof-path tiles (Queue/Graph/Reports ->
  Findings/Security Graph/Compliance).
